Fix MeshType format warning and guard SDL2 imgui backend
- Cast enum class MeshType to int for %d format specifier (fixes -Wformat error on Linux)
- Guard dear_imgui SDL2 backend with if(TARGET SDL2::SDL2-static) so builds without SDL2 succeed

+++ src/experimental/filament/filament/model_objects.cc
@@ -415,7 +415,7 @@ static std::span<const int> GetIndices(const mjModel* model,
 static void UpdateMeshData(MeshData* data, const mjModel* model, int id,
                            MeshType mesh_type) {
   if (!IsValidIndex(model, id, mesh_type)) {
-    mju_error("Invalid index %d for type %d", id, mesh_type);
+    mju_error("Invalid index %d for type %d", id, static_cast<int>(mesh_type));
     return;
   }
